wire 수정
