+++ iot_gateway/src/main/java/com/nhnacademy/node/MqttInNode.java
@@ -1,0 +1,27 @@
+package com.nhnacademy.node;
+
+public class MqttInNode extends InputNode {
+
+    MqttInNode(int count) {
+        super(count);
+    }
+
+    MqttInNode() {
+        this(1);
+    }
+
+    @Override
+    void preprocess() {
+
+    }
+
+    @Override
+    void process() {
+
+    }
+
+    @Override
+    void postprocess() {
+
+    }
+}
