--- iot_gateway/src/main/java/com/nhnacademy/wire/BufferedWire.java
@@ -1,27 +1,0 @@
-package com.nhnacademy.wire;
-
-import java.util.LinkedList;
-import java.util.Queue;
-
-import com.nhnacademy.message.Message;
-
-public class BufferedWire implements Wire {
-    Queue<Message> messageQueue;
-
-    public BufferedWire() {
-        super();
-        messageQueue = new LinkedList<>();
-    }
-
-    public void put(Message message) {
-        messageQueue.add(message);
-    }
-
-    public boolean hasMessage() {
-        return !messageQueue.isEmpty();
-    }
-
-    public Message get() {
-        return messageQueue.poll();
-    }
-}
